markers are circles
