+++ pages/dev/test-site/explore/main.styl
@@ -187,7 +187,7 @@ h3,h4,p
   margin-right 10px
 
 .marker_pin
-  background-image url(https://storage.macrostrat.org/assets/rockd/marker_red.png)
+  background-image url(https://storage.macrostrat.org/assets/rockd/explore/red-circle.png)
   width 60px
   height 60px
   background-size 50%
@@ -197,7 +197,7 @@ h3,h4,p
 
   span
     position absolute
-    top 45%
+    top 50%
     left 50%
     transform translate(-50%, -50%)
     color white
